+++ lib/utils.js
@@ -10,4 +10,15 @@ function basename (filename) {
   return dotPosition === -1 ? filename : filename.substring(0, dotPosition)
 }
 
-module.exports = { basename }
+/**
+ * Check if a given string is a valid string
+ *
+ * @param sString - the string that needs to be checked
+ */
+function isString (sString) {
+  if (typeof (sString) !== 'string' || sString === '') {
+    throw new Error(`${sString} is not a valid string`)
+  }
+}
+
+module.exports = { basename, isString }
